bugfix: fix infinite loop
Calling `_send_request_to_node` with the result of `least_loaded_node`
results in an infinite loop if the metadata is out-dated and
`least_loaded_node` return the `node_id` of a broker that was removed
from the cluster. The method `_send_request_to_least_loaded_node`
handles this race and fixes the infinite loop.